refactor: lift the cloud dashboard projection to shared components
projectCloudNotebookDashboard and its types lived in apps/notebook-cloud/viewer,
while the Elements catalog re-implemented its own projection beside it. Move the
pure projection to src/components/notebook/workspace/notebook-dashboard.ts so
cloud and Elements import the same code; cloud importers now resolve it through
@/components/notebook/workspace. Pure code move, no behavior change.
